+++ CHANGELOG.md
@@ -3,5 +3,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+0.0.2
+  * add GeoJSONTileLayer plugin
+  * add TileLayer.Ajax plugin
+  * add TopoTileLayer plugin
+  * add CanvasLayer plugin
+  * add Feature adapter
+
 0.0.1
   * initial release